hyperionrelay/requirements.txt
certifi==2018.1.18
chardet==3.0.4
cryptography==1.7.1
idna==2.6
keyring==10.1
keyrings.alt==1.3
protobuf==3.5.2
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
pyserial==3.4
python-apt==1.1.0b5
pyxdg==0.25
requests==2.18.4
RPi.GPIO==0.6.3
SecretStorage==2.3.1
six==1.11.0
urllib3==1.22
